Add Docker support for xml-fixer
This commit introduces Docker support to the xml-fixer project by adding a .dockerignore file to exclude unnecessary files from the Docker build context, a docker-compose.yml file for defining and running multi-container Docker applications, and a Dockerfile to build a Docker image for the application. These additions facilitate the dockerization of the xml-fixer application, making it easier to deploy and run in isolated environments.

Add docker getting started

Added docker documentation If somebody prefer to use Docker for running the `xml-fixer`